+++ src/commonMain/kotlin/teksturepako/pakku/api/actions/export/rules/CurseForgeRules.kt
@@ -67,7 +67,7 @@ fun createCfModpackModel(
 
         CfModLoaderData(
             id = "$loaderName-$loaderVersion",
-            primary = lockFile.getLoadersWithVersions().firstOrNull()!! == loader
+            primary = lockFile.getLoadersWithVersions().firstOrNull() == loader
         )
     }