+++ src/commonMain/kotlin/teksturepako/pakku/api/data/PAKKU_DIR.kt
@@ -1,2 +1,0 @@
-package teksturepako.pakku.api.data
-