+++ src/commonMain/kotlin/teksturepako/pakku/cli/cmd/Fetch.kt
@@ -4,29 +4,21 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.terminal
 import com.github.ajalt.mordant.animation.progressAnimation
 import com.github.ajalt.mordant.widgets.Spinner
-import com.github.michaelbull.result.get
 import com.github.michaelbull.result.getOrElse
-import com.github.michaelbull.result.runCatching
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import teksturepako.pakku.api.actions.ActionError.AlreadyExists
+import teksturepako.pakku.api.actions.fetch.deleteOldFiles
 import teksturepako.pakku.api.actions.fetch.fetch
 import teksturepako.pakku.api.actions.fetch.retrieveProjectFiles
-import teksturepako.pakku.api.actions.sync.getFileHashes
 import teksturepako.pakku.api.actions.sync.sync
 import teksturepako.pakku.api.data.LockFile
-import teksturepako.pakku.api.data.workingPath
 import teksturepako.pakku.api.overrides.readProjectOverrides
 import teksturepako.pakku.api.platforms.Multiplatform
-import teksturepako.pakku.api.projects.ProjectType
 import teksturepako.pakku.cli.ui.pDanger
 import teksturepako.pakku.cli.ui.pError
 import teksturepako.pakku.cli.ui.pInfo
 import teksturepako.pakku.cli.ui.pSuccess
-import teksturepako.pakku.io.createHash
-import kotlin.io.path.*
 
 class Fetch : CliktCommand("Fetch projects to your modpack folder")
 {
@@ -51,72 +43,56 @@ class Fetch : CliktCommand("Fetch projects to your modpack folder")
             }
         }
 
-        projectFiles.fetch(
-            onError = { error ->
-                if (error !is AlreadyExists) terminal.pError(error)
-            },
-            onProgress = { advance, total, sent ->
-                progressBar.advance(advance)
-                progressBar.updateTotal(total)
+        val fetchJob = launch {
+            projectFiles.fetch(
+                onError = { error ->
+                    if (error !is AlreadyExists) terminal.pError(error)
+                },
+                onProgress = { advance, total ->
+                    progressBar.advance(advance)
+                    progressBar.updateTotal(total)
+                },
+                onSuccess = { projectFile, _ ->
+                    terminal.pSuccess("${projectFile.getPath()} saved")
+                },
+                lockFile
+            )
+        }
 
-                if (sent >= total)
-                {
-                    progressBar.clear()
-                    echo()
-                }
-            },
-            onSuccess = { projectFile, _ ->
-                terminal.pSuccess("${projectFile.getPath()} saved")
-            },
-            lockFile
-        )
+        fetchJob.invokeOnCompletion {
+            progressBar.clear()
+        }
 
         // -- OVERRIDES --
 
         val projectOverrides = readProjectOverrides()
 
-        projectOverrides.sync(
-            onError = { error ->
-                if (error !is AlreadyExists) terminal.pError(error)
-            },
-            onSuccess = { projectOverride ->
-                terminal.pInfo("${projectOverride.fullOutputPath} synced")
-            }
-        ).joinAll()
+        val syncJob = launch {
+            projectOverrides.sync(
+                onError = { error ->
+                    if (error !is AlreadyExists) terminal.pError(error)
+                },
+                onSuccess = { projectOverride ->
+                    terminal.pInfo("${projectOverride.fullOutputPath} synced")
+                }
+            )
+        }
 
         // -- OLD FILES --
 
-        val oldFiles = ProjectType.entries
-            .filterNot { it == ProjectType.WORLD }
-            .mapNotNull { projectType ->
-                val folder = Path(workingPath, projectType.folderName)
-                if (folder.notExists()) return@mapNotNull null
-                runCatching { folder.listDirectoryEntries() }.get()
-            }.flatMap { entry ->
-                entry.filter { file ->
-                    val bytes = runCatching { file.readBytes() }.get()
-                    val fileHash = bytes?.let { createHash("sha1", it) }
+        val oldFilesJob = launch {
+            deleteOldFiles(
+                onSuccess = { file ->
+                    terminal.pDanger("$file deleted")
+                },
+                projectFiles, projectOverrides
+            )
+        }
 
-                    val projectFileNames = projectFiles.filter { projectFile ->
-                        projectFile.hashes?.get("sha1") == null
-                    }.map { projectFile ->
-                        projectFile.fileName
-                    }
+        fetchJob.join()
+        syncJob.join()
+        oldFilesJob.join()
 
-                    file.extension in listOf("jar", "zip")
-                            && fileHash !in projectOverrides.getFileHashes()
-                            && file.name !in projectFileNames
-                            && fileHash !in projectFiles.mapNotNull { projectFile ->
-                                projectFile.hashes?.get("sha1")
-                            }
-                }
-            }
-
-        oldFiles.map {
-            launch(Dispatchers.IO) {
-                it.deleteIfExists()
-                terminal.pDanger("$it deleted")
-            }
-        }.joinAll()
+        echo()
     }
 }